Change the percentage of time I aim to spend on development tasks
Recently I measured the time I spend on various tasks. It came out as:
- 25% management
- 40% development
- 35% "stuff" :)

I still have to figure out what "stuff" is, but it turns out that I spend quite a bit of my time on technical chat - more than I realised. So I've updated this to reflect that more clearly.